Implement an inline method-target cache.
Add code to the prologue of every ccbuf to check whether %eax is as
expected, and jump out to bt_continue to do a frag_hash lookup if
not. This allows us to chain calls to JMP targets, effectively
implementing an inline cache for JMP calls.

The additional check at the start of every prologue causes a small
overall performance penalty. Whether this effect or the inline cache
is dominant depends on the test case, but future work should remove
this unnecessary check in many common cases.