Skip the (PC == %eax) check in most frags.
Generate the pc-checking prologue for every frag, but store a pointer
immediately after the check as the code buffer in the frag cache,
causing us to skip it most of the time. Give bt_translate_and_run an
additional 'exact' parameter, which indicates whether it should chain
to the start of a cfrag, or to the PC-checking prologue, and add a new
bt_continue_ic, which is like bt_continue_chain but unsets the 'exact'
flag.